pyproject.toml
[tool.poetry]
name = "ai-governance-firewall"
version = "0.1.0"
description = "AI Governance Firewall — API proxy for LLM governance, detection, and compliance"
authors = ["AI Governance Team"]
readme = "README.md"
packages = [
    { include = "proxy" },
    { include = "detection" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.34.0" }
httpx = "^0.28.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
structlog = "^24.4.0"
redis = { extras = ["hiredis"], version = "^5.2.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.36" }
asyncpg = "^0.30.0"
alembic = "^1.14.0"
spacy = "^3.8.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
python-multipart = "^0.0.19"
prometheus-client = "^0.21.0"
orjson = "^3.10.0"
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.25.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
locust = "^2.32.0"
testcontainers = { extras = ["postgres", "redis"], version = "^4.9.0" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
